+++ src/App.css
@@ -4,6 +4,20 @@ body {
   background-color: #004080;
 }
 
+.linkTo {
+  display: block;
+  width: 100%;
+  height: 100%;
+  color: white;
+  text-decoration: none;
+  padding: 20px;
+  display: flex;
+  box-sizing: border-box;
+  flex-direction: column;
+  justify-content: space-between;
+  text-align: center;
+}
+
 .backView {
   transition: all 0.5s ease;
   position: absolute;
@@ -232,14 +246,8 @@ header {
 }
 
 .instrumentPanel {
-  display: flex;
-  box-sizing: border-box;
-  flex-direction: column;
-  justify-content: space-between;
-  text-align: center;
   width: 25%;
   min-width: 280px;
-  padding: 20px;
   font-family: 'Lato', sans-serif;
   color: white;
   border: 1px solid rgba(100, 100, 100, 0.2);
@@ -278,6 +286,7 @@ header {
   font-family: Lato, Arial, sans-serif;
   font-size: 18px;
   font-weight: 700;
+  text-decoration: none;
   padding: 10px 20px 10px 50px;
   cursor: pointer;
   margin-right: 20px;
