+++ src/api/instruments.js
@@ -1,2119 +1,15 @@
 import { initInstruments } from '../actions/actionsIndex';
-// import { appAlert } from '../utils';
+import { appAlert } from '../utils';
 
-// async function downloadInstruments(store) {
-//   document.body.style.cursor = 'wait';
-//   const instrumentsURL = 'https://financialmodelingprep.com/api/v3/stock/list?apikey=2d204314b27b541f159ab9757c0b8c78';
-//   fetch(instrumentsURL)
-//     .then(response => response.json())
-//     .then(data => {
-//       const sortedInstruments = data.sort((a, b) => ((a.symbol < b.symbol) ? -1 : 1));
-//       store.dispatch(initInstruments(sortedInstruments));
-//     })
-//     .catch(err => appAlert('downloadInstruments', 'Error :  '.concat(err)));
-// }
 async function downloadInstruments(store) {
-  store.dispatch(initInstruments([
-    {
-      symbol: 'SPY',
-      name: 'SPDR S&amp;P 500',
-      price: 337.12,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'CMCSA',
-      name: 'Comcast Corp',
-      price: 46.35,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'KMI',
-      name: 'Kinder Morgan Inc',
-      price: 12.23,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'INTC',
-      name: 'Intel Corp',
-      price: 52.31,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'MU',
-      name: 'Micron Technology Inc',
-      price: 47.755,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'GDX',
-      name: 'VanEck Vectors Gold Miners',
-      price: 39.75,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'GE',
-      name: 'General Electric Co',
-      price: 6.155,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'BAC',
-      name: 'Bank of America Corp',
-      price: 24.02,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'EEM',
-      name: 'iShares MSCI Emerging Index Fund',
-      price: 44.36,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'XLF',
-      name: 'SPDR Select Sector Fund - Financial',
-      price: 24.105,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'AAPL',
-      name: 'Apple Inc',
-      price: 116.44,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'MSFT',
-      name: 'Microsoft Corp',
-      price: 212.53,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'SIRI',
-      name: 'Sirius XM Holdings Inc',
-      price: 5.525,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'HPQ',
-      name: 'HP Inc',
-      price: 19.085,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'CX',
-      name: 'Cemex S.A.B. de C.V. Sponsored ADR',
-      price: 3.68,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'EFA',
-      name: 'iShares MSCI EAFE',
-      price: 63.995,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'CZR',
-      name: 'Caesars Entertainment Inc',
-      price: 56.23,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'QQQ',
-      name: 'PowerShares QQQ Trust Series 1',
-      price: 281.3,
-      exchange: 'NASDAQ Global Market',
-    },
-    {
-      symbol: 'F',
-      name: 'Ford Motor Co',
-      price: 6.715,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'AMD',
-      name: 'Advanced Micro Devices Inc',
-      price: 83.7689,
-      exchange: 'NASDAQ Capital Market',
-    },
-    {
-      symbol: 'SNAP',
-      name: 'Snap Inc',
-      price: 26.545,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'FB',
-      name: 'Facebook Inc',
-      price: 267.205,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'WFC',
-      name: 'Wells Fargo &amp; Co',
-      price: 23.25,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'AIG',
-      name: 'American International Group Inc',
-      price: 27.685,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'T',
-      name: 'AT&amp;T Inc',
-      price: 28.5473,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'C',
-      name: 'Citigroup Inc',
-      price: 42.79,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'VALE',
-      name: 'VALE S.A. American Depositary Shares Each Representing one',
-      price: 10.335,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'MS',
-      name: 'Morgan Stanley',
-      price: 47.87,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'AKS',
-      name: 'AK Steel Holding Corporation',
-      price: 1.55,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'JPM',
-      name: 'JPMorgan Chase &amp; Co',
-      price: 96.6802,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'ORCL',
-      name: 'Oracle Corp',
-      price: 59.9899,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'NKE',
-      name: 'Nike Inc',
-      price: 126.448,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'VWO',
-      name: 'Vanguard FTSE Emerging Markets',
-      price: 43.49,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'PG',
-      name: 'Procter &amp; Gamble Co',
-      price: 139.81,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'FXI',
-      name: 'iShares China Large-Cap',
-      price: 42.385,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'IWM',
-      name: 'iShares Russell 2000',
-      price: 150.28,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'GSM',
-      name: 'Ferroglobe PLC',
-      price: 0.6411,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'HK',
-      name: 'Halcon Resources Corporation',
-      price: 0.19,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'BBD',
-      name: 'Banco Bradesco Sa American Depositary Shares',
-      price: 3.39,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'PFE',
-      name: 'Pfizer Inc',
-      price: 36.375,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'JD',
-      name: 'JD.com Inc.',
-      price: 78.495,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'NOK',
-      name: 'Nokia Corporation Sponsored American Depositary Shares',
-      price: 3.86,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'TWTR',
-      name: 'Twitter Inc',
-      price: 45.62,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'AMLP',
-      name: 'Alerian MLP',
-      price: 19.71,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'AVGO',
-      name: 'Broadcom Inc',
-      price: 366.81,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'AMAT',
-      name: 'Applied Materials Inc',
-      price: 60.779,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'WFT',
-      name: 'Weatherford International plc (Ireland)',
-      price: 0.36,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'UVXY',
-      name: 'ProShares Trust Ultra VIX Short Term Futures',
-      price: 20.19,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'DWDP',
-      name: 'DowDuPont Inc.',
-      price: 30.51,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'VXX',
-      name: 'iPath S&amp;P 500 VIX Short Term Futures TM ETN',
-      price: 25.1216,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'VEA',
-      name: 'Vanguard FTSE Developed Markets',
-      price: 41.1361,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'ZNGA',
-      name: 'Zynga Inc',
-      price: 9.46,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'XOM',
-      name: 'Exxon Mobil Corp',
-      price: 33.245,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'QCOM',
-      name: 'Qualcomm Inc',
-      price: 118.493,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'TVIX',
-      name: 'VelocityShares Daily 2x VIX Short Term ETN',
-      price: 112.36,
-      exchange: 'NASDAQ Global Market',
-    },
-    {
-      symbol: 'VIPS',
-      name: 'Vipshop Holdings Ltd',
-      price: 15.818,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'GLD',
-      name: 'SPDR Gold Trust',
-      price: 179.37,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'CSCO',
-      name: 'Cisco Systems Inc',
-      price: 38.939,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'AXP',
-      name: 'American Express Co',
-      price: 100.367,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'BMY',
-      name: 'Bristol-Myers Squibb Co',
-      price: 59.9805,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'V',
-      name: 'Visa Inc',
-      price: 204.18,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'USO',
-      name: 'United States Oil Fund',
-      price: 27.09,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'GRPN',
-      name: 'Groupon Inc',
-      price: 21.04,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'WP',
-      name: 'Worldpay Inc. Class A',
-      price: 124.37,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'OIH',
-      name: 'VanEck Vectors Oil Services',
-      price: 93.94,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'GERN',
-      name: 'Geron Corp',
-      price: 1.75,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'KEY',
-      name: 'KeyCorp',
-      price: 11.87,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'RF',
-      name: 'Regions Financial Corp',
-      price: 11.435,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'KR',
-      name: 'Kroger Co',
-      price: 34.149,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'HAL',
-      name: 'Halliburton Co',
-      price: 11.295,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'BABA',
-      name: 'Alibaba Group Holding Ltd',
-      price: 291.299,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'MRO',
-      name: 'Marathon Oil Corp',
-      price: 3.955,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'CLNS',
-      name: 'Colony NorthStar Inc.',
-      price: 6.41,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'GILD',
-      name: 'Gilead Sciences Inc',
-      price: 62.9,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'IEMG',
-      name: 'iShares Core MSCI Emerging Markets',
-      price: 53.105,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'GM',
-      name: 'General Motors Co',
-      price: 30.385,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'FCX',
-      name: 'Freeport-McMoRan Inc',
-      price: 15.645,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'CRM',
-      name: 'Salesforce.Com Inc',
-      price: 253.48,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'ATVI',
-      name: 'Activision Blizzard Inc',
-      price: 82.81,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'SQ',
-      name: 'Square Inc',
-      price: 166.44,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'P',
-      name: 'Pandora Media Inc.',
-      price: 8.4,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'XLK',
-      name: 'SPDR Select Sector Fund - Technology',
-      price: 117.8,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'TWX',
-      name: 'Time Warner Inc.',
-      price: 98.91,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'NUE',
-      name: 'Nucor Corp',
-      price: 45.17,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'XOP',
-      name: 'SPDR S&amp;P Oil &amp; Gas Explor &amp; Product',
-      price: 40.72,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'SWN',
-      name: 'Southwestern Energy Co',
-      price: 2.245,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'LOW',
-      name: 'Lowe`s Companies Inc',
-      price: 167.48,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'RAD',
-      name: 'Rite Aid Corp',
-      price: 9.33,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'VEON',
-      name: 'VEON Ltd.',
-      price: 1.26,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'HYG',
-      name: 'iShares iBoxx $ High Yield Corporate Bond',
-      price: 83.865,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'APC',
-      name: 'Anadarko Petroleum Corporation',
-      price: 69.92,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'JNK',
-      name: 'SPDR Bloomberg Barclays High Yield Bond',
-      price: 104.27,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'EWJ',
-      name: 'iShares MSCI Japan Index Fund',
-      price: 59.09,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'XLU',
-      name: 'SPDR Select Sector Fund - Utilities',
-      price: 60,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'ESV',
-      name: 'Ensco plc Class A',
-      price: 7.38,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'SLB',
-      name: 'Schlumberger NV',
-      price: 15.155,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'FLEX',
-      name: 'Flex Ltd',
-      price: 11.41,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'FOXA',
-      name: 'Fox Corp',
-      price: 27.68,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'ABBV',
-      name: 'Abbvie Inc',
-      price: 87.29,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'GIS',
-      name: 'General Mills Inc',
-      price: 61.99,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'VZ',
-      name: 'Verizon Communications Inc',
-      price: 59.37,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'XRX',
-      name: 'Xerox Holdings Corp',
-      price: 18.69,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'CVLT',
-      name: 'Commvault Systems Inc',
-      price: 41.04,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'IEFA',
-      name: 'iShares Core MSCI EAFE',
-      price: 60.64,
-      exchange: 'BATS Exchange',
-    },
-    {
-      symbol: 'X',
-      name: 'United States Steel Corp',
-      price: 7.4819,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'MAT',
-      name: 'Mattel Inc',
-      price: 11.9,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'CTL',
-      name: 'Centurylink Inc',
-      price: 11,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'MIK',
-      name: 'Michaels Companies Inc',
-      price: 10.385,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'DVN',
-      name: 'Devon Energy Corp',
-      price: 9.3,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'BKLN',
-      name: 'PowerShares Exchange-Traded Fund Trust',
-      price: 21.8,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'IBN',
-      name: 'ICICI Bank Limited',
-      price: 10.21,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'MPC',
-      name: 'Marathon Petroleum Corp',
-      price: 27.78,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'EZU',
-      name: 'iShares MSCI Eurozone',
-      price: 38.18,
-      exchange: 'BATS Exchange',
-    },
-    {
-      symbol: 'PM',
-      name: 'Philip Morris International Inc',
-      price: 74.03,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'MSCI',
-      name: 'MSCI Inc',
-      price: 361.33,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'RDN',
-      name: 'Radian Group Inc',
-      price: 14.825,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'KO',
-      name: 'Coca-Cola Co',
-      price: 49.092,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'AABA',
-      name: 'Altaba Inc.',
-      price: 19.63,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'SRC',
-      name: 'Spirit Realty Capital Inc',
-      price: 34.135,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'NFX',
-      name: 'Newfield Exploration Company',
-      price: 17.03,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'COP',
-      name: 'ConocoPhillips',
-      price: 32.21,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'CHK',
-      name: 'Chesapeake Energy Corporation',
-      price: 11.85,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'DIS',
-      name: 'Walt Disney Co',
-      price: 123.74,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'LEN',
-      name: 'Lennar Corp',
-      price: 82.2101,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'CVX',
-      name: 'Chevron Corp',
-      price: 70.945,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'PYPL',
-      name: 'PayPal Holdings Inc',
-      price: 197.98,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'GDXJ',
-      name: 'VanEck Vectors Junior Gold Miners',
-      price: 56.64,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'CTRP',
-      name: 'Ctrip.com International Ltd.',
-      price: 35.97,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'AUY',
-      name: 'Yamana Gold Inc',
-      price: 5.805,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'DBEF',
-      name: 'Xtrackers MSCI EAFE Hedged Equity',
-      price: 30.69,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'WMT',
-      name: 'Walmart Inc',
-      price: 143.11,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'CLF',
-      name: 'Cleveland-Cliffs Inc',
-      price: 6.7,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'MRK',
-      name: 'Merck &amp; Co Inc',
-      price: 82.1,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'MDLZ',
-      name: 'Mondelez International Inc',
-      price: 57.72,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'RSPP',
-      name: 'RSP Permian Inc.',
-      price: 47.93,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'GG',
-      name: 'Goldcorp Inc.',
-      price: 11.22,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'MGM',
-      name: 'MGM Resorts International',
-      price: 21.795,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'GFI',
-      name: 'Gold Fields Limited American Depositary Shares',
-      price: 12.43,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'KGC',
-      name: 'Kinross Gold Corp',
-      price: 9.08,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'SCHW',
-      name: 'Charles Schwab Corp',
-      price: 37.245,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'NBR',
-      name: 'Nabors Industries Ltd',
-      price: 23.455,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'SRCI',
-      name: 'SRC Energy Inc.',
-      price: 4,
-      exchange: 'NYSE American',
-    },
-    {
-      symbol: 'JAG',
-      name: 'Jagged Peak Energy Inc.',
-      price: 8.23,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'NVDA',
-      name: 'NVIDIA Corp',
-      price: 542.768,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'JNJ',
-      name: 'Johnson &amp; Johnson',
-      price: 148.261,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'WDC',
-      name: 'Western Digital Corp',
-      price: 36.228,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'BSX',
-      name: 'Boston Scientific Corp',
-      price: 38.64,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'STI',
-      name: 'SunTrust Banks Inc.',
-      price: 70.13,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'ON',
-      name: 'ON Semiconductor Corp',
-      price: 22.805,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'ITUB',
-      name: 'Itau Unibanco Banco Holding SA American Depositary Shares (Each repstg 500 Preferred shares)',
-      price: 3.895,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'BK',
-      name: 'Bank of New York Mellon Corp',
-      price: 34.02,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'HMY',
-      name: 'Harmony Gold Mining Company Limited',
-      price: 5.37,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'APA',
-      name: 'Apache Corp',
-      price: 9.17,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'HOME',
-      name: 'At Home Group Inc',
-      price: 15.18,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'NGD',
-      name: 'New Gold Inc',
-      price: 1.795,
-      exchange: 'NYSE American',
-    },
-    {
-      symbol: 'KOS',
-      name: 'Kosmos Energy Ltd',
-      price: 0.9642,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'DAL',
-      name: 'Delta Air Lines Inc',
-      price: 30.95,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'INFY',
-      name: 'Infosys Limited American Depositary Shares',
-      price: 13.755,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'CCL',
-      name: 'Carnival Corp',
-      price: 14.82,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'BHGE',
-      name: 'Baker Hughes a GE company Class A',
-      price: 25.18,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'MFGP',
-      name: 'Micro Focus Intl PLC ADS each representing One Ord Sh',
-      price: 3.2,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'PBR',
-      name: 'Petroleo Brasileiro S.A.- Petrobras',
-      price: 6.855,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'DISCA',
-      name: 'Discovery Inc',
-      price: 21.35,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'LPI',
-      name: 'Laredo Petroleum Inc',
-      price: 9.19,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'ABT',
-      name: 'Abbott Laboratories',
-      price: 109.59,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'FITB',
-      name: 'Fifth Third Bancorp',
-      price: 20.985,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'PTEN',
-      name: 'Patterson-UTI Energy Inc',
-      price: 2.61,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'USB',
-      name: 'U.S. Bancorp',
-      price: 35.74,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'BA',
-      name: 'Boeing Co',
-      price: 167.789,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'GGB',
-      name: 'Gerdau S.A.',
-      price: 3.585,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'JBLU',
-      name: 'JetBlue Airways Corp',
-      price: 11.43,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'QEP',
-      name: 'QEP Resources Inc',
-      price: 0.888,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'EBAY',
-      name: 'eBay Inc',
-      price: 52.6,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'HBAN',
-      name: 'Huntington Bancshares Inc',
-      price: 9.145,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'DB',
-      name: 'Deutsche Bank AG',
-      price: 8.39,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'MRVL',
-      name: 'Marvell Technology Group Ltd',
-      price: 40.1,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'ABX',
-      name: 'Barrick Gold Corporation',
-      price: 13.55,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'WBA',
-      name: 'Walgreens Boots Alliance Inc',
-      price: 35.89,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'CDNA',
-      name: 'CareDx Inc',
-      price: 41.18,
-      exchange: 'NASDAQ Global Market',
-    },
-    {
-      symbol: 'ETP',
-      name: 'Energy Transfer Partners L.P. representing limited partner interests',
-      price: 21.49,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'RSX',
-      name: 'VanEck Vectors Russia',
-      price: 20.845,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'KNX',
-      name: 'Knight-Swift Transportation Holdings Inc',
-      price: 40.22,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'KBH',
-      name: 'KB Home',
-      price: 38.8096,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'DISCK',
-      name: 'Discovery Communications Inc.',
-      price: 19.23,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'PGR',
-      name: 'Progressive Corp',
-      price: 94.13,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'XLE',
-      name: 'SPDR Select Sector Fund - Energy Select Sector',
-      price: 29.1319,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'MO',
-      name: 'Altria Group Inc',
-      price: 38.71,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'MDT',
-      name: 'Medtronic PLC',
-      price: 103.21,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'RIG',
-      name: 'Transocean Ltd',
-      price: 0.782,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'BCS',
-      name: 'Barclays PLC',
-      price: 4.99,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'CDEV',
-      name: 'Centennial Resource Development Inc',
-      price: 0.5795,
-      exchange: 'NASDAQ Capital Market',
-    },
-    {
-      symbol: 'DNR',
-      name: 'Denbury Resources Inc',
-      price: 0.241,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'EOG',
-      name: 'EOG Resources Inc',
-      price: 34.91,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'XLP',
-      name: 'SPDR Select Sector Fund - Consumer Staples',
-      price: 64.44,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'HPE',
-      name: 'Hewlett Packard Enterprise Co',
-      price: 9.285,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'TXN',
-      name: 'Texas Instruments Inc',
-      price: 145,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'OAS',
-      name: 'Oasis Petroleum Inc',
-      price: 0.2152,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'OKE',
-      name: 'ONEOK Inc',
-      price: 25.61,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'CVS',
-      name: 'CVS Health Corp',
-      price: 57.735,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'SVXY',
-      name: 'ProShares Short VIX Short Term Futures',
-      price: 35.33,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'HES',
-      name: 'Hess Corp',
-      price: 39.3,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'GDDY',
-      name: 'GoDaddy Inc',
-      price: 77.135,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'BX',
-      name: 'Blackstone Group Inc',
-      price: 53.32,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'NLSN',
-      name: 'Nielsen Holdings PLC',
-      price: 14.295,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'TS',
-      name: 'Tenaris S.A. American Depositary Shares',
-      price: 9.655,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'BOX',
-      name: 'Box Inc',
-      price: 17.64,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'OXY',
-      name: 'Occidental Petroleum Corp',
-      price: 9.675,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'OCLR',
-      name: 'Oclaro Inc.',
-      price: 8.27,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'CVE',
-      name: 'Cenovus Energy Inc',
-      price: 3.61,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'SYF',
-      name: 'Synchrony Financial',
-      price: 26.405,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'IAU',
-      name: 'ishares Gold Trust',
-      price: 18.225,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'MTG',
-      name: 'MGIC Investment Corp',
-      price: 9.045,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'LUV',
-      name: 'Southwest Airlines Co',
-      price: 37.865,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'FTI',
-      name: 'TechnipFMC PLC',
-      price: 6.16,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'AA',
-      name: 'Alcoa Corp',
-      price: 11.52,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'ABEV',
-      name: 'Ambev S.A. American Depositary Shares (Each representing 1)',
-      price: 2.225,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'VTI',
-      name: 'Vanguard Total Stock Market',
-      price: 171.515,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'EQR',
-      name: 'Equity Residential',
-      price: 52.55,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'EWG',
-      name: 'iShares MSCI Germany Index Fund',
-      price: 29.28,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'ESRX',
-      name: 'Express Scripts Holding Company',
-      price: 92.48,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'MDR',
-      name: 'McDermott International Inc.',
-      price: 78,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'ETSY',
-      name: 'ETSY Inc',
-      price: 129.28,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'ANDV',
-      name: 'Andeavor',
-      price: 152.35,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'TAL',
-      name: 'TAL Education Group',
-      price: 77.14,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'UNH',
-      name: 'UnitedHealth Group Inc',
-      price: 311.81,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'ECA',
-      name: 'Encana Corporation',
-      price: 3.79,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'GES',
-      name: 'Guess? Inc',
-      price: 11.99,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'AMZN',
-      name: 'Amazon.com Inc',
-      price: 3199.05,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'NFLX',
-      name: 'Netflix Inc',
-      price: 510.89,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'ENTG',
-      name: 'Entegris Inc',
-      price: 75.8074,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'CSX',
-      name: 'CSX Corp',
-      price: 77.63,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'WYNN',
-      name: 'Wynn Resorts Ltd',
-      price: 71.3801,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'FOX',
-      name: 'Twenty-First Century Fox Inc.',
-      price: 27.85,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'CNC',
-      name: 'Centene Corp',
-      price: 57.9,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'ZION',
-      name: 'Zions Bancorporation NA',
-      price: 28.99,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'CLDR',
-      name: 'Cloudera Inc',
-      price: 10.85,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'RES',
-      name: 'RPC Inc',
-      price: 2.635,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'CI',
-      name: 'Cigna Corp',
-      price: 167.31,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'KRE',
-      name: 'SPDR S&amp;P Regional Banking',
-      price: 35.67,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'WMB',
-      name: 'Williams Companies Inc',
-      price: 19.02,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'SYNH',
-      name: 'Syneos Health Inc',
-      price: 55.6131,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'HBI',
-      name: 'HanesBrands Inc',
-      price: 15.75,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'BBT',
-      name: 'BB&amp;T Corporation',
-      price: 54.24,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'XL',
-      name: 'XL Group Ltd.',
-      price: 57.59,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'PAYX',
-      name: 'Paychex Inc',
-      price: 79.54,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'CELG',
-      name: 'Celgene Corporation',
-      price: 108.24,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'D',
-      name: 'Dominion Energy Inc',
-      price: 79.3,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'CHS',
-      name: 'Chico\'s FAS Inc',
-      price: 0.9514,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'CFG',
-      name: 'Citizens Financial Group Inc',
-      price: 25.345,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'XLI',
-      name: 'SPDR Select Sector Fund - Industrial',
-      price: 77.215,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'SU',
-      name: 'Suncor Energy Inc',
-      price: 11.655,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'SLV',
-      name: 'iShares Silver Trust',
-      price: 22.44,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'MMM',
-      name: '3M Co',
-      price: 159.768,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'BIIB',
-      name: 'Biogen Inc',
-      price: 281.645,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'JBL',
-      name: 'Jabil Inc',
-      price: 34.9,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'ACN',
-      name: 'Accenture PLC',
-      price: 226.627,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'NTR',
-      name: 'Nutrien Ltd',
-      price: 38.17,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'PAH',
-      name: 'Platform Specialty Products Corporation',
-      price: 11.26,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'BRFS',
-      name: 'BRF S.A.',
-      price: 3.185,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'SBUX',
-      name: 'Starbucks Corp',
-      price: 87.2,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'EWU',
-      name: 'iShares MSCI United Kingdom',
-      price: 25.53,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'PE',
-      name: 'Parsley Energy Inc',
-      price: 9.175,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'LVS',
-      name: 'Las Vegas Sands Corp',
-      price: 46.19,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'EGO',
-      name: 'Eldorado Gold Corp',
-      price: 10.655,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'GSK',
-      name: 'GlaxoSmithKline PLC',
-      price: 37.29,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'TEVA',
-      name: 'Teva Pharmaceutical Industries Limited American Depositary Shares',
-      price: 9.025,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'PAA',
-      name: 'Plains All American Pipeline LP',
-      price: 5.89,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'SQQQ',
-      name: 'ProShares UltraPro Short QQQ',
-      price: 22.985,
-      exchange: 'NASDAQ Global Market',
-    },
-    {
-      symbol: 'VLO',
-      name: 'Valero Energy Corp',
-      price: 39.825,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'CSRA',
-      name: 'CSRA Inc.',
-      price: 41.23,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'IMAX',
-      name: 'Imax Corp',
-      price: 11.69,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'FNB',
-      name: 'F.N.B. Corp',
-      price: 6.73,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'ADSK',
-      name: 'Autodesk Inc',
-      price: 235.89,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'HD',
-      name: 'Home Depot Inc',
-      price: 278.04,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'BIDU',
-      name: 'Baidu Inc',
-      price: 126.01,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'AR',
-      name: 'Antero Resources Corp',
-      price: 2.705,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'ARNC',
-      name: 'Arconic Corp (PITTSBURGH)',
-      price: 19.07,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'VOYA',
-      name: 'Voya Financial Inc',
-      price: 48.25,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'SPXU',
-      name: 'ProShares UltraPro Short S&amp;P500',
-      price: 8.54,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'MLCO',
-      name: 'Melco Resorts &amp; Entertainment Ltd',
-      price: 16.49,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'SJNK',
-      name: 'SPDR Bloomberg Barclays Short Term High Yield Bond',
-      price: 25.835,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'CWB',
-      name: 'SPDR Bloomberg Barclays Convertible Securities',
-      price: 69.63,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'SM',
-      name: 'SM Energy Co',
-      price: 1.5,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'NTRS',
-      name: 'Northern Trust Corp',
-      price: 77.2,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'DHI',
-      name: 'D.R. Horton Inc',
-      price: 76.41,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'UAL',
-      name: 'United Airlines Holdings Inc',
-      price: 34.97,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'HRB',
-      name: 'H &amp; R Block Inc',
-      price: 16.525,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'BTU',
-      name: 'Peabody Energy Corp',
-      price: 2.225,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'JBGS',
-      name: 'JBG SMITH Properties',
-      price: 26.91,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'IPG',
-      name: 'Interpublic Group of Companies Inc',
-      price: 16.775,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'DE',
-      name: 'Deere &amp; Co',
-      price: 221.78,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'KHC',
-      name: 'Kraft Heinz Co',
-      price: 30.02,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'COG',
-      name: 'Cabot Oil &amp; Gas Corp',
-      price: 17.055,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'IBM',
-      name: 'International Business Machines Corp',
-      price: 121.9,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'WTW',
-      name: 'Weight Watchers International Inc',
-      price: 19.04,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'NYT',
-      name: 'New York Times Co',
-      price: 43.07,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'DBJP',
-      name: 'Xtrackers MSCI Japan Hedged Equity',
-      price: 40.66,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'AAL',
-      name: 'American Airlines Group Inc',
-      price: 12.495,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'OLN',
-      name: 'Olin Corp',
-      price: 12.5766,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'JCI',
-      name: 'Johnson Controls International PLC',
-      price: 40.99,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'VICI',
-      name: 'VICI Properties Inc',
-      price: 23.46,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'VST',
-      name: 'Vistra Corp',
-      price: 18.155,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'NWL',
-      name: 'Newell Brands Inc',
-      price: 17.055,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'EPD',
-      name: 'Enterprise Products Partners LP',
-      price: 15.755,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'KMB',
-      name: 'Kimberly-Clark Corp',
-      price: 148.62,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'AEP',
-      name: 'American Electric Power Company Inc',
-      price: 82.54,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'IT',
-      name: 'Gartner Inc',
-      price: 125.455,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'CAT',
-      name: 'Caterpillar Inc',
-      price: 148.5,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'AU',
-      name: 'AngloGold Ashanti Limited',
-      price: 26.6,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'MYL',
-      name: 'Mylan NV',
-      price: 14.87,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'XRAY',
-      name: 'DENTSPLY SIRONA Inc',
-      price: 43.27,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'KKR',
-      name: 'KKR &amp; Co Inc',
-      price: 34.69,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'IEUR',
-      name: 'iShares Core MSCI Europe',
-      price: 45.04,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'LQD',
-      name: 'iShares iBoxx $ Investment Grade Corporate Bond',
-      price: 134.53,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'CMC',
-      name: 'Commercial Metals Co',
-      price: 20.02,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'JCP',
-      name: 'J.C. Penney Company Inc. Holding Company',
-      price: 0.1819,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'XBI',
-      name: 'SPDR Series Trust S&amp;P Biotech',
-      price: 111.24,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'SYMC',
-      name: 'Symantec Corporation',
-      price: 19.98,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'TMUS',
-      name: 'T-Mobile US Inc',
-      price: 115.085,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'EWW',
-      name: 'iShares Inc MSCI Mexico',
-      price: 33.1,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'RMP',
-      name: 'Rice Midstream Partners LP representing limited partner interests',
-      price: 18.14,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'NRG',
-      name: 'NRG Energy Inc',
-      price: 30.73,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'CPB',
-      name: 'Campbell Soup Co',
-      price: 48.47,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'ATHM',
-      name: 'Autohome Inc',
-      price: 97.73,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'WRB',
-      name: 'W. R. Berkley Corp',
-      price: 60.93,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'MOMO',
-      name: 'Momo Inc',
-      price: 14.5399,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'TGT',
-      name: 'Target Corp',
-      price: 159.93,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'TV',
-      name: 'Grupo Televisa S.A.',
-      price: 6.205,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'GLW',
-      name: 'Corning Inc',
-      price: 32.35,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'MAS',
-      name: 'Masco Corp',
-      price: 55.52,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'DOV',
-      name: 'Dover Corp',
-      price: 108.03,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'HRL',
-      name: 'Hormel Foods Corp',
-      price: 49.06,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'S',
-      name: 'Sprint Corporation',
-      price: 8.62,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'TECK',
-      name: 'Teck Resources Ltd',
-      price: 14.04,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'AES',
-      name: 'AES Corp',
-      price: 17.94,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'PPL',
-      name: 'PPL Corp',
-      price: 27.59,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'IXUS',
-      name: 'iShares Core MSCI Total International Stock ETF',
-      price: 58.68,
-      exchange: 'NASDAQ Global Market',
-    },
-    {
-      symbol: 'SEE',
-      name: 'Sealed Air Corp',
-      price: 39.64,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'NLY',
-      name: 'Annaly Capital Management Inc',
-      price: 7.0476,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'FRC',
-      name: 'First Republic Bank',
-      price: 109.92,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'EWT',
-      name: 'iShares Inc MSCI Taiwan',
-      price: 45.43,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'STX',
-      name: 'Seagate Technology PLC',
-      price: 49.24,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'ADBE',
-      name: 'Adobe Inc',
-      price: 497.9,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'K',
-      name: 'Kellogg Co',
-      price: 64.21,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'RTN',
-      name: 'Raytheon Company',
-      price: 116.96,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'M',
-      name: 'Macy\'s Inc',
-      price: 5.815,
-      exchange: 'New York Stock Exchange',
-    },
-    {
-      symbol: 'TSLA',
-      name: 'Tesla Inc',
-      price: 439.126,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'CY',
-      name: 'Cypress Semiconductor Corporation',
-      price: 23.82,
-      exchange: 'Nasdaq Global Select',
-    },
-    {
-      symbol: 'KBE',
-      name: 'SPDR S&amp;P Bank',
-      price: 29.47,
-      exchange: 'NYSE Arca',
-    },
-    {
-      symbol: 'TSM',
-      name: 'Taiwan Semiconductor Manufacturing Company Ltd.',
-      price: 82.24,
-      exchange: 'New York Stock Exchange',
-    },
-  ]));
+  document.body.style.cursor = 'wait';
+  const instrumentsURL = 'https://financialmodelingprep.com/api/v3/stock/list?apikey=2d204314b27b541f159ab9757c0b8c78';
+  fetch(instrumentsURL)
+    .then(response => response.json())
+    .then(data => {
+      const sortedInstruments = data.sort((a, b) => ((a.symbol < b.symbol) ? -1 : 1));
+      store.dispatch(initInstruments(sortedInstruments));
+    })
+    .catch(err => appAlert('downloadInstruments', 'Error :  '.concat(err)));
 }
 export default downloadInstruments;
