+++ src/containers/instrumentProfile.js
@@ -1,19 +1,4 @@
-/* eslint-disable jsx-a11y/click-events-have-key-events */
-/* eslint-disable jsx-a11y/no-static-element-interactions */
-
-import React, { useEffect } from 'react';
-import { connect } from 'react-redux';
-import PropTypes from 'prop-types';
-import downloadProfile from '../api/profile';
-import { gel, appAlert } from '../utils';
-
-function loadItemData(itemId, data) {
-  try {
-    gel(itemId).innerHTML = data;
-  } catch (error) {
-    appAlert(itemId, error);
-  }
-}
+import React from 'react';
 
 function renderItem(label, itemId) {
   return (
@@ -26,52 +11,7 @@ function renderItem(label, itemId) {
   );
 }
 
-function profileCallBack(data) {
-  const sector = data.sector || '';
-  const industry = data.industry || '';
-  const state = data.state || '';
-  const country = data.country || '';
-  const city = data.city || '';
-  const address = data.address || '';
-  const zip = data.zip || '';
-
-  const companyAddress = country
-    .concat(' : ')
-    .concat(state)
-    .concat(', ')
-    .concat(city)
-    .concat(', ')
-    .concat(address)
-    .concat(', zip ')
-    .concat(zip);
-
-  loadItemData('companySector', sector.concat(' - ').concat(industry));
-  loadItemData('companyName', data.companyName);
-  loadItemData('companyAddress', companyAddress);
-  loadItemData('companyDescription', data.description);
-  loadItemData('companyWebsite', data.website);
-  loadItemData('companyPhone', data.phone);
-  loadItemData('companyExchange', data.exchange);
-  loadItemData('companyEmployees', data.fullTimeEmployees);
-  loadItemData('companyCEO', data.ceo);
-  loadItemData('companyVolume', data.volAvg);
-  loadItemData('companyCapital', data.mktCap);
-
-  try {
-    gel('companyLogo').src = data.image;
-  } catch (error) {
-    appAlert('companyLogo', error);
-  }
-}
-
-function InstrumentProfile(props) {
-  const { instrument } = props;
-  const { symbol } = instrument;
-
-  useEffect(() => {
-    downloadProfile(symbol, profileCallBack);
-  });
-
+function InstrumentProfile() {
   return (
     <div className="instrumentProfile">
       {renderItem('Industry Sector', 'companySector')}
@@ -94,12 +34,4 @@ function InstrumentProfile(props) {
   );
 }
 
-InstrumentProfile.propTypes = {
-  instrument: PropTypes.objectOf(PropTypes.any),
-};
-
-InstrumentProfile.defaultProps = {
-  instrument: null,
-};
-
-export default connect(null, null)(InstrumentProfile);
+export default InstrumentProfile;
