+++ src/containers/instrumentsListHeader.js
@@ -7,22 +7,6 @@ import ExchangeFilter from '../components/exchangeFilter';
 function InstrumentsListHeader(props) {
   const { exchangesData, filterChanged } = props;
   const { exchanges, filter } = exchangesData;
-  // const { exchangesData, filterChanged } = props;
-  // const { exchanges, filter } = exchangesData;
-  // const exchangeOptions = [];
-  // exchangeOptions.push(
-  //   <option key={0} value="All Exchanges">All Exchanges</option>,
-  // );
-  // let index = 1;
-  // exchanges.map(exchangeyName => {
-  //   index += 1;
-  //   return exchangeOptions.push(
-  //     <option key={index} value={exchangeyName}>
-  //       {exchangeyName}
-  //     </option>,
-  //   );
-  // });
-
   return (
     <header>
       <div className="headerContainer">
@@ -38,13 +22,6 @@ function InstrumentsListHeader(props) {
             filter={filter}
           />
         </div>
-        {/* <div className="headerContainer-right">
-          <div className="exchangeContainer">
-            <select name="exchange" id="exchange" value={filter} onChange={filterChanged}>
-              {exchangeOptions}
-            </select>
-          </div>
-        </div> */}
       </div>
     </header>
   );
